requirements.txt
asgiref==3.8.1
astroid==3.3.9
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
idna==3.10
isort==6.0.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
numpy==2.2.4
openapi-codec==1.3.2
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
psycopg2-binary==2.9.10
pylint==3.3.6
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.3
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
